SSCS-7515-7516 Latest refactor
