+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/WriteFinalDecisionAboutToSubmitHandler.java
@@ -77,7 +77,8 @@ public class WriteFinalDecisionAboutToSubmitHandler implements PreSubmitCallback
 
     private <T extends PointsCondition<?>> List<String> getDecisionNoticePointsValidationErrorMessages(Class<T> enumType, DecisionNoticeQuestionService decisionNoticeQuestionService, SscsCaseData sscsCaseData) {
         return Arrays.stream(enumType.getEnumConstants())
-            .filter(pointsCondition -> pointsCondition.isApplicable(sscsCaseData))
+            .filter(pointsCondition -> pointsCondition.isApplicable(
+                decisionNoticeQuestionService, sscsCaseData))
             .map(pointsCondition ->
                 pointsCondition.getOptionalErrorMessage(decisionNoticeQuestionService, sscsCaseData))
             .filter(Optional::isPresent)
