+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/writefinaldecision/pip/PipPointsCondition.java
@@ -1,12 +1,18 @@
 package uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.pip;
 
+import static org.apache.commons.collections4.CollectionUtils.emptyIfNull;
+
+import java.util.Collection;
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Function;
 import java.util.function.IntPredicate;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
+import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.ActivityAnswer;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.ActivityType;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.AwardType;
 import uk.gov.hmcts.reform.sscs.ccd.presubmit.writefinaldecision.PointsCondition;
+import uk.gov.hmcts.reform.sscs.service.DecisionNoticeQuestionService;
 
 /**
  * Enum encapsulating the attributes of a points-related condition on SscsCaseData. Each condition specifies the type of award the condition applies for, the activity type it applies to, along with
@@ -46,7 +52,7 @@ public enum PipPointsCondition implements PointsCondition<PipPointsCondition> {
         this.errorMessage = getStandardErrorMessage(awardType, activityType);
     }
 
-    public boolean isApplicable(SscsCaseData sscsCaseData) {
+    public boolean isApplicable(DecisionNoticeQuestionService questionService, SscsCaseData sscsCaseData) {
         return awardType.getKey().equals(activityType.getAwardTypeExtractor().apply(sscsCaseData));
     }
 
@@ -86,5 +92,27 @@ public enum PipPointsCondition implements PointsCondition<PipPointsCondition> {
         }
         return "You have previously selected " + awardDescription + " for " + activityType.getName()
             + ". The points awarded don't match. Please review your previous selection.";
+    }
+
+    /**
+     * Given a points condition, and an SscsCaseData instance, obtain an error message for that condition if the condition has failed to be satified, or an empty optional if the condition is met.
+     *
+     * @param questionService The question service to use
+     * @param sscsCaseData The SscsCaseData to evaluate against the condition.
+     * @return An optional error message if the condition has failed to be satified, or an empty optional if the condition is met.
+     */
+    public Optional<String> getOptionalErrorMessage(DecisionNoticeQuestionService questionService, SscsCaseData sscsCaseData) {
+
+        Collection<String> collection = emptyIfNull(getAnswersExtractor().apply(sscsCaseData));
+
+        if (collection.isEmpty()) {
+            return Optional.empty();
+        }
+        int totalPoints = collection.stream().map(answerText -> questionService.getAnswerForActivityQuestionKey(sscsCaseData,
+            answerText)).filter(Optional::isPresent).map(Optional::get).mapToInt(ActivityAnswer::getActivityAnswerPoints).sum();
+
+        return getPointsRequirementCondition().test(totalPoints) ? Optional.empty() :
+            Optional.of(getErrorMessage());
+
     }
 }
